Change default ProtVer from 1337 to 14
See Art-Net 4 Protocol Release V1.4 Document Revision 1.4dd 22/1/2017 page 47.
Also see
https://art-net.org.uk/structure/streaming-packets/artdmx-packet-definition/
under ProtVer.

Increasing ProtVer above 14 works today. But in the future a device may
actually support a higher version of Art-Net and then we may be
breaking compatibility because we don't support this higher version but
claim that we do.
It's a funny easter egg, but breaking forward compatibility should
be avoided.